refactor(devsettings): move DeveloperSettingsScreen to impl, delegate via BlocScreen
DeveloperSettingsScreen moves to developer-settings/impl/ui.
DeveloperSettingsBloc extends BlocScreen; the inline preview fake
in the screen file gains a Content() override.
RootBloc.Child.DeveloperSettings gets `BlocScreen by bloc`.

developer-settings/public sets publicResClass = true;
developer-settings/impl gains the Compose plugin + ui/text deps.
