+++ client/developer-settings/impl/build.gradle.kts
@@ -1,14 +1,21 @@
-plugins { alias(libs.plugins.kmpLibrary) }
+plugins {
+    alias(libs.plugins.kmpLibrary)
+    alias(libs.plugins.compose)
+}
 
 kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(projects.client.developerSettings.public)
+            implementation(projects.client.text.public)
+            implementation(projects.client.ui.public)
             implementation(libs.arkivanov.decompose.core)
+            implementation(libs.arkivanov.decompose.compose.extensions)
             implementation(projects.client.shared)
             implementation(projects.client.auth.data.public)
             implementation(projects.client.auth.usecase.public)
             implementation(projects.client.recipe.data.public)
+            implementation(compose.components.resources)
         }
         commonTest.dependencies { implementation(projects.client.auth.data.testing) }
     }
