+++ client/developer-settings/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/devsettings/impl/DeveloperSettingsBlocImpl.kt
@@ -1,11 +1,14 @@
 package com.plusmobileapps.chefmate.devsettings.impl
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.Environment
 import com.plusmobileapps.chefmate.devsettings.DeveloperSettingsBloc
 import com.plusmobileapps.chefmate.devsettings.DeveloperSettingsBloc.Output
 import com.plusmobileapps.chefmate.devsettings.TestUser
+import com.plusmobileapps.chefmate.devsettings.impl.ui.DeveloperSettingsScreen
 import com.plusmobileapps.chefmate.di.AppScope
 import com.plusmobileapps.chefmate.getViewModel
 import com.plusmobileapps.metro.extensions.assistedfactory.ContributesAssistedFactory
@@ -71,5 +74,10 @@ class DeveloperSettingsBlocImpl(
 
     override fun onFeatureFlagsClicked() {
         output.onNext(Output.OpenFeatureFlags)
+    }
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        DeveloperSettingsScreen(bloc = this, modifier = modifier)
     }
 }
