+++ client/developer-settings/public/build.gradle.kts
@@ -15,4 +15,6 @@ kotlin {
     }
 }
 
+compose { resources { publicResClass = true } }
+
 plusLibrary { namespace = "com.plusmobileapps.chefmate.devsettings" }
